refactor: simulation configuration document model
