--- src/models/db/base_model.py
@@ -1,26 +1,0 @@
-from datetime import datetime
-
-from pydantic import BaseModel, Field
-from bson.objectid import ObjectId as BsonObjectId
-
-
-class PydanticObjectId(BsonObjectId):
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, v):
-        if not isinstance(v, BsonObjectId):
-            raise TypeError('ObjectId required')
-        return v
-
-
-class MetadataBaseDoc(BaseModel):
-    id: PydanticObjectId = Field(..., alias='_id')
-    inserted_at: datetime = datetime.utcnow()
-    updated_at: datetime = datetime.utcnow()
-
-    class Config:
-        allow_population_by_field_name = True
-        extra = 'forbid'
